pyproject.toml
[project]
name = "qq-manager"
version = "0.1.0-dev.1"
description = "Wrapper around PBS Pro and Slurm batch systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
    "readchar>=4.2.1",
    "rich>=14.1.0",
]

[dependency-groups]
dev = ["nuitka>=2.7.16", "ruff>=0.13.0", "ty>=0.0.1a20"]

[tool.ruff.lint]
extend-select = [
    "F",   # Pyflakes rules
    "W",   # PyCodeStyle warnings
    "E",   # PyCodeStyle errors
    "I",   # Sort imports properly
    "UP",  # Warn if certain things can changed due to newer Python versions
    "C4",  # Catch incorrect use of comprehensions, dict, list, etc
    "FA",  # Enforce from __future__ import annotations
    "ISC", # Good use of string concatenation
    "ICN", # Use common import conventions
    "RET", # Good return practices
    "SIM", # Common simplification rules
    "TID", # Some good import practices
    "TC",  # Enforce importing certain types in a TYPE_CHECKING block
    "PTH", # Use pathlib instead of os.path
    "TD",  # Be diligent with TODO comments
    "NPY", # Some numpy-specific things
    "ARG", # Catch incorrect use of arguments
]
ignore = ["E501"]
